# specs/openapi/member-service.yaml
openapi: 3.0.1
info:
  title: Member Service API
  description: 회원 관리, 상품 관리 및 배송지 관리 API 명세
  version: 0.0.3
servers:
- url: http://localhost:8080
paths:
  /members:
    get:
      tags:
      - members
      operationId: members-find-all
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
    post:
      tags:
      - members
      operationId: members-join
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberJoinRequest'
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
  /members/{id}:
    get:
      tags:
      - members
      operationId: members-find-by-id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
    put:
      tags:
      - members
      operationId: members-update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdateRequest'
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
    delete:
      tags:
      - members
      operationId: members-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
  /members/{id}/addresses:
    get:
      tags:
      - addresses
      summary: 회원의 배송지 목록 조회
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressResponse'
    post:
      tags:
      - addresses
      summary: 배송지 신규 등록
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
  /members/{id}/addresses/{addressId}:
    put:
      tags:
      - addresses
      summary: 배송지 정보 수정
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: addressId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
    delete:
      tags:
      - addresses
      summary: 배송지 삭제
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: addressId
        in: path
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: 성공
  /products:
    get:
      tags:
      - products
      operationId: products-find-all
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    post:
      tags:
      - products
      operationId: products-register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRegisterRequest'
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
  /products/{id}:
    get:
      tags:
      - products
      operationId: products-find-by-id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    put:
      tags:
      - products
      operationId: products-update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    delete:
      tags:
      - products
      operationId: products-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
components:
  schemas:
    MemberUpdateRequest:
      type: object
      properties:
        name:
          type: string
    MemberResponse:
      type: array
      items:
        type: object
        properties:
          id:
            type: number
          email:
            type: string
          name:
            type: string
          role:
            type: string
    MemberJoinRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    AddressRequest:
      type: object
      properties:
        nickname:
          type: string
          description: 배송지 별명
        address:
          type: string
          description: 상세 주소
        zipCode:
          type: string
          description: 우편번호
    AddressResponse:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        address:
          type: string
        zipCode:
          type: string
    ProductUpdateRequest:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        stock:
          type: number
        category:
          type: string
        brandId:
          type: number
    ProductRegisterRequest:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        stock:
          type: number
        category:
          type: string
        brandId:
          type: number
    ProductResponse:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        price:
          type: number
        stock:
          type: number
        category:
          type: string
        brandName:
          type: string